make test-programs crate responsible for running tests
like upstream, i'm skipping all proc macros and just generating whole
source files in build.rs. This doesn't make any difference except
to put all the code in one place.
