+++ host/Cargo.toml
@@ -13,11 +13,3 @@ wasmtime-wasi-sockets-sync = { workspace = true }
 clap = { version = "4.1.9", features = ["derive"] }
 tracing-subscriber = { version = "0.3", default-features = false, features = ["env-filter", "fmt" ]}
 tokio = { version = "1", features = ["full"] }
-
-[dev-dependencies]
-tracing = { workspace = true }
-cap-rand = { workspace = true }
-test-programs = { path = "../test-programs" }
-test-log = { version = "0.2", default-features = false, features = ["trace"] }
-tempfile = "3.3.0"
-lazy_static = "1"
